requirements.txt
django
djangorestframework
djoser
djangorestframework-simplejwt
